added video component
