+++ src/components/pages/home/home.js
@@ -3,11 +3,12 @@ import { fontStyles, fontSizing, spacing, colors, containerSizing } from '../../
 import { ENDPOINTS } from '../../../resources/constants';
 import { VimeoService } from '../../../services/vimeoService';
 import './home.css';
+import { Video } from '../../common/video';
 
 const sectionStyles = {
     display: 'flex',
     flexDirection: 'column',
-    justifyContent: 'center',
+    justifyContent: 'space-around',
     alignItems: 'center'
 };
 
@@ -15,7 +16,8 @@ const divStyles = {
     display: 'flex',
     flexDirection: 'row',
     justifyContent: 'space-around',
-    alignItems: 'center'
+    alignItems: 'center',
+    flexWrap: 'wrap',
 };
 
 const hStyles = {
@@ -34,15 +36,6 @@ const h2Styles = {
     margin: spacing.small,
     fontSize: fontSizing.medium,
     ...hStyles
-}; 
-
-const cardStyles = {
-    maxWidth: containerSizing.medium,
-    boxShadow: colors.cardBoxShadow,
-    borderRadius: 7,
-    margin: spacing.medium,
-    backgroundColor: colors.cardBackgroundColor,
-    objectFit: 'cover'
 };
 
 const vimeoService = new VimeoService();
@@ -65,7 +58,7 @@ export default class Home extends React.Component {
             console.log('techniques', this.state.techniques);
         } catch (exception) {
             console.log('EXCEPTIONNN', exception);
-        }        
+        }
     }
 
     render() {
@@ -75,13 +68,11 @@ export default class Home extends React.Component {
                 <h1 style={h1Styles}>Mahecha BJJ</h1>
                 <h2 style={h2Styles}>Technique Taught to the Point</h2>
                 <div style={divStyles}>
-                {this.state.techniques.map(technique => {
-                    return (
-                        <video key={technique.name} style={cardStyles} src={technique.files[1].link} controls controlsList="nodownload">
-                        </video>
-                    );
-                    
-                })}
+                    {this.state.techniques.map(technique => {
+                        return (
+                            <Video technique={technique} />
+                        );
+                    })}
                 </div>
             </section>
         );
